Deleted app route to draw_map and fixed update bug
The update bug where the data was not updating properly has been fixed. Old data is now summed with the new submission, which creates the most up-to-date and accurate total. This fix revealed a new bug where if the map was reloaded it would add the new submission's data to the map, this was also fixed. App route to draw_map was deleted. New route to heatmap.html will be added soon.